Module 2/backend/requirements.txt
# Backend Requirements
# Module 2: CV Parsing Module
# Note: Optimized for Python 3.14 with pre-built wheels

# FastAPI & Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
motor>=3.3.0
pymongo>=4.5.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# NLP & Text Processing
spacy>=3.7.0
nltk>=3.8.0
PyPDF2>=3.0.0
python-docx>=0.8.11

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Security
PyJWT>=2.8.0
python-jose>=3.3.0
bcrypt>=4.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
